CLOG-152 Fix BlogWow Import
These changes fix the BlogWow import. They add the clog tool to a site
if the site does not have the clog tool (this is necessary as
a check is made to see if the tool is in the site later on).

Also adds in a couple workarounds for issues noticed when importing:
* Checks to see if a site's group's names are all present
* Also properly escapes ' in the title of posts